+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Fixed `pagemedia` field not uploading/deleting files right away
     * Fixed `Flex Pages` add, copy and move buttons appearing in edit view when no permissions
     * Fixed `Flex Pages` permission issues
+    * Fixed some admin redirect issues
 
 # v1.0.0-rc.12
 ## 06/08/2020